Extract `TestClassRegistrar`
