+++ common/junit-platform-native/src/main/java/org/graalvm/junit/platform/JUnitPlatformFeature.java
@@ -66,11 +66,9 @@ import java.io.InputStreamReader;
 import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.ServiceLoader;
-import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
@@ -157,6 +155,12 @@ public final class JUnitPlatformFeature implements Feature {
                 .selectors(selectors)
                 .build();
 
+        TestClassRegistrar testClassRegistrar = new TestClassRegistrar(clazz -> {
+            debug("Registering test class for reflection: %s", clazz.getName());
+            nativeImageConfigImpl.registerAllClassMembersForReflection(clazz);
+            forEachProvider(p -> p.onTestClassRegistered(clazz, nativeImageConfigImpl));
+        });
+
         Launcher launcher = LauncherFactory.create();
         TestPlan testPlan = launcher.discover(request);
         testPlan.getRoots().stream()
@@ -167,49 +171,7 @@ public final class JUnitPlatformFeature implements Feature {
                 .filter(ClassSource.class::isInstance)
                 .map(ClassSource.class::cast)
                 .map(ClassSource::getJavaClass)
-                .forEach(this::registerTestClassForReflection);
-    }
-
-    private final Set<Class<?>> registeredClasses = new HashSet<>();
-
-    private boolean shouldRegisterClass(Class<?> clazz) {
-        /* avoid registering java internal classes */
-        if (ModuleLayer.boot().modules().contains(clazz.getModule())) {
-            return false;
-        }
-
-        /* avoid loops (possible case: class B is inner class of A, and B extends A) */
-        if (registeredClasses.contains(clazz)) {
-            return false;
-        }
-        registeredClasses.add(clazz);
-
-        return true;
-    }
-
-    private void registerTestClassForReflection(Class<?> clazz) {
-        if (!shouldRegisterClass(clazz)) {
-            return;
-        }
-
-        debug("Registering test class for reflection: %s", clazz.getName());
-        nativeImageConfigImpl.registerAllClassMembersForReflection(clazz);
-        forEachProvider(p -> p.onTestClassRegistered(clazz, nativeImageConfigImpl));
-
-        Class<?>[] declaredClasses = clazz.getDeclaredClasses();
-        for (Class<?> declaredClass : declaredClasses) {
-            registerTestClassForReflection(declaredClass);
-        }
-
-        Class<?>[] interfaces = clazz.getInterfaces();
-        for (Class<?> inter : interfaces) {
-            registerTestClassForReflection(inter);
-        }
-
-        Class<?> superClass = clazz.getSuperclass();
-        if (superClass != null && superClass != Object.class) {
-            registerTestClassForReflection(superClass);
-        }
+                .forEach(testClassRegistrar::registerTestClassForReflection);
     }
 
     private void forEachProvider(Consumer<PluginConfigProvider> consumer) {
